fix: fix memory leaking in CodeEditor
